+++ docs/src/app/pages/docs/components/app-components-inline-editor/app-components-inline-editor.tsx
@@ -43,6 +43,7 @@ export class AppComponentsInlineEditor {
 <li><a href="#app-components-inline-editor-properties">Properties</a></li>
 <li><a href="#app-components-inline-editor-styling">Styling</a></li>
 <li><a href="#app-components-inline-editor-events">Events</a></li>
+<li><a href="#app-components-inline-editor-methods">Methods</a></li>
 <li><a href="#app-components-inline-editor-examples">Examples</a></li>
 </ul>
 </li>
@@ -70,11 +71,11 @@ export class AppComponentsInlineEditor {
     </deckgo-highlight-code><h3 id="app-components-inline-editor-install-from-npm">Install from NPM</h3>
 <p>Install it in your project from <a href="https://www.npmjs.com/package/@deckdeckgo/qrcode">npm</a> using the following command:</p>
 <deckgo-highlight-code language="bash">
-      <code slot="code">npm install deckdeckgo-inline-editor</code>
+      <code slot="code">npm install @deckdeckgo&#47;inline-editor</code>
     </deckgo-highlight-code><h3 id="app-components-inline-editor-framework-integration">Framework integration</h3>
 <p>The <a href="https://stenciljs.com/docs/overview">Stencil documentation</a> provide examples of framework integration for <a href="https://stenciljs.com/docs/angular">Angular</a>, <a href="https://stenciljs.com/docs/react">React</a>, <a href="https://stenciljs.com/docs/vue">Vue</a> and <a href="https://stenciljs.com/docs/ember">Ember</a>.</p>
 <h2 id="app-components-inline-editor-usage">Usage</h2>
-<p>The <code>&lt;deckgo-inline-editor/&gt;</code> should be added once only in your page. It will interact with all elements of types <code>p</code>, <code>h1</code>, <code>h2</code>  and <code>h3</code> you would have set as <code>contenteditable</code>.</p>
+<p>The <code>&lt;deckgo-inline-editor/&gt;</code> should be added once only in your page. It will interact with all elements of types <code>p</code>, <code>h1</code>, <code>h2</code>  and <code>h3</code>, or other <code>containers</code> you would define, which are set as <code>contenteditable</code>.</p>
 <h3 id="app-components-inline-editor-properties">Properties</h3>
 <p>The <code>&lt;deckgo-inline-editor/&gt;</code> expose the following properties:</p>
 <table>
@@ -104,7 +105,7 @@ export class AppComponentsInlineEditor {
 <tr>
 <td><code>stickyMobile</code></td>
 <td><code>sticky-mobile</code></td>
-<td>Use a sticky footer toolbar on mobile. Note: except iOS, feel free to send a PR if you know how to handle this with the software keyboard</td>
+<td>Use a sticky footer toolbar on mobile. The sticky bar is positioned bottom except on iOS for which it will be positioned top.</td>
 <td><code>boolean</code></td>
 <td><code>false</code></td>
 </tr>
@@ -172,6 +173,11 @@ export class AppComponentsInlineEditor {
 <td>--deckgo-inline-editor-sticky-bottom</td>
 <td>0</td>
 <td>The bottom attribute of the sticky toolbar</td>
+</tr>
+<tr>
+<td>--deckgo-inline-editor-sticky-zindex</td>
+<td></td>
+<td>The z-Index of the sticky toolbar</td>
 </tr>
 <tr>
 <td>--deckgo-inline-editor-separator-background</td>
@@ -295,8 +301,17 @@ export class AppComponentsInlineEditor {
 <td>Triggered when an image is manipulated. Note: the event won&#39;t provide directly the image but rather its container element.</td>
 <td><code>CustomEvent&lt;HTMLElement&gt;</code></td>
 </tr>
+<tr>
+<td><code>stickyToolbarActivated</code></td>
+<td>Triggered when the sticky toolbar would be activated or not. Useful for example if you want to catch the event to hide things in your footer, as the sticky toolbar is display above it.</td>
+<td><code>CustomEvent&lt;boolean&gt;</code></td>
+</tr>
 </tbody></table>
-<h3 id="app-components-inline-editor-examples">Examples</h3>
+<h3 id="app-components-inline-editor-methods">Methods</h3>
+<p>This component also export an async method <code>reset()</code> which will reset the inline editor (= hide it) and optionally clear its selection.</p>
+<deckgo-highlight-code language="javascript">
+      <code slot="code">const element = document.querySelector(&#039;deckgo-inline-editor&#039;);{'\n'}await element.reset(clearSelection: boolean, blurActiveElement?: boolean);</code>
+    </deckgo-highlight-code><h3 id="app-components-inline-editor-examples">Examples</h3>
 <deckgo-highlight-code language="javascript">
       <code slot="code">&lt;p contenteditable slot=&quot;content&quot;&gt;This text will be editable&lt;&#47;p&gt;{'\n'}{'\n'}&lt;h1 contenteditable slot=&quot;title&quot;&gt;This title too&lt;&#47;h1&gt;{'\n'}{'\n'}&lt;deckgo-inline-editor&gt;&lt;&#47;deckgo-inline-editor&gt;</code>
     </deckgo-highlight-code></main>