+++ remote/src/components.d.ts
@@ -7,9 +7,9 @@
 
 import '@stencil/core';
 
+import '@deckdeckgo/core';
 import '@ionic/core';
 import 'ionicons';
-import '@deckdeckgo/core';
 import {
   DeckdeckgoSlideDefinition,
 } from '@deckdeckgo/types';