+++ studio/src/app/components/feed/app-popular/app-popular.tsx
@@ -26,6 +26,16 @@ export class AppPopular {
                 </ion-card-content>
 
                 <app-demo></app-demo>
+            </ion-card>,
+            <h1 class="ion-padding ion-margin-start ion-margin-end">We need your help!</h1>,
+            <ion-card>
+                <ion-card-content>
+                    <p>We are in development, we need your help.</p>
+
+                    <p class="ion-padding-top">Send us your feedback via <a href="mailto:hello@deckdeckgo.com">email</a>, <a href="https://twitter.com/deckdeckgo">Twitter</a> or on <a href="https://join.slack.com/t/deckdeckgo/shared_invite/enQtNjMyNTk2NTQwODk5LTAxZjAwZWQwODQyZDg1ZDA5ODhlOTE3OGMwZjhmYjY3NDRhZjViZTRiNWU3OGU3MjYyNjE1OWE3NzNkZmQ3ZWI" target="_blank">Slack</a>.</p>
+
+                    <p class="ion-padding-top">You are awesome!</p>
+                </ion-card-content>
             </ion-card>
         ];
     }