+++ studio/src/app/definitions/resize-observer.d.ts
@@ -1,0 +1,11 @@
+interface ResizeObserverConstructor {
+    new (callback: any);
+
+    observe: (target: HTMLElement) => void;
+
+    unobserve: (target: HTMLElement) => void;
+
+    disconnect: () => void;
+}
+
+declare var ResizeObserver: ResizeObserverConstructor;