+++ webcomponents/core/src/components/slides/deckdeckgo-slide-youtube/deckdeckgo-slide-youtube.tsx
@@ -1,7 +1,9 @@
 import {Component, Element, Event, EventEmitter, Method, Prop, State} from '@stencil/core';
 
+import {DeckDeckGoUtils} from '@deckdeckgo/utils';
+
 import {DeckdeckgoSlide, DeckdeckgoSlideUtils} from '../deckdeckgo-slide';
-import {DeckdeckgoUtils} from '../../utils/deckdeckgo-utils';
+import {DeckdeckgoDeckUtils} from '../../utils/deckdeckgo-deck-utils';
 
 @Component({
   tag: 'deckgo-slide-youtube',
@@ -29,7 +31,7 @@ export class DeckdeckgoSlideYoutube implements DeckdeckgoSlide {
   @Prop({reflectToAttr: true}) customBackground: boolean = false;
 
   async componentDidLoad() {
-    await DeckdeckgoUtils.hideLazyLoadImages(this.el);
+    await DeckdeckgoDeckUtils.hideLazyLoadImages(this.el);
 
     this.initWindowResize();
 
@@ -137,7 +139,7 @@ export class DeckdeckgoSlideYoutube implements DeckdeckgoSlide {
 
   private initWindowResize() {
     if (window) {
-      window.addEventListener('resize', DeckdeckgoUtils.debounce(this.onResizeContent));
+      window.addEventListener('resize', DeckDeckGoUtils.debounce(this.onResizeContent));
     }
   }